requirements.txt
# Core web
fastapi==0.115.0
uvicorn==0.30.0

# NLP / ML
spacy==3.7.2
nltk==3.9
transformers==4.40.0
scikit-learn==1.5.0
joblib==1.3.2

# Speech / TTS
SpeechRecognition==3.10.0
pyaudio==0.2.14
gTTS==2.5.1
faster-whisper==0.6.0
pyttsx3==2.90

# OpenAI and async file helpers
openai==1.0.0
aiofiles==23.1.0
python-multipart==0.0.6

# Database / ORM
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
alembic==1.13.1

# Utilities
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.0
pydantic==2.7.0
werkzeug==3.1.5

# Security / Auth
passlib[bcrypt]==1.7.4
python-jose==3.3.0

# Testing
pytest==8.3.0
pytest-asyncio==0.23.0
coverage==7.6.0

# Extras / Tools
loguru==0.7.2
pipwin==0.5.1

# Data / Analysis
pandas==2.2.2
matplotlib==3.9.2
seaborn==0.13.2
streamlit==1.41.1
openpyxl==3.1.5
XlsxWriter==3.2.0
fpdf==1.7.2

# Media / FFmpeg helpers
imageio-ffmpeg==0.5.1
ffmpeg-python==0.2.0
moviepy==1.0.3

# RTL / complex scripts
arabic-reshaper==2.1.4
python-bidi==0.4.2
